requirements.txt
abydos==0.5.0
aiohttp==3.7.4.post0
anyio==3.6.2
async-generator==1.10
async-timeout==3.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
decorator==5.1.1
deprecation==2.1.0
edgartools==1.15.5
exceptiongroup==1.1.1
fastcore==1.5.29
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
humanize==4.6.0
idna==3.4
jellyfish==0.11.2
Levenshtein==0.21.0
lxml==4.9.2
markdown-it-py==2.2.0
markdownify==0.11.6
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
py==1.11.0
pyarrow==12.0.0
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.21.0
PythonTurtle==0.3.2
pytz==2023.3
pyxDamerauLevenshtein==1.7.1
rank-bm25==0.2.1
rapidfuzz==3.0.0
requests==2.30.0
retry==0.9.2
rich==13.3.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
textdistance==4.5.0
tk==0.1.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wsproto==1.2.0
yarl==1.9.2
